pyproject.toml
[tool.poetry]
name = "healthygamersearchengine"
version = "0.1.0"
description = ""
authors = ["lil-sussy <naykeysnet@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "==3.12.2"
django = "^5.0.4"
djangorestframework = "^3.15.1"
requests = "^2.31.0"
chromadb = "^0.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
